services/patient_service/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
mysqlclient>=2.1.0
django-environ>=0.10.0
drf-yasg==1.21.5
pyyaml>=6.0
uritemplate>=4.1.1
django-cors-headers>=4.3.1 